moved nickname cycle
