+++ .env.tpl
@@ -3,3 +3,5 @@ TOKEN=""
 GIF_TOKEN=""
 GOOGLE_APPLICATION_CREDENTIALS_JSON=''
 GOOGLE_APPLICATION_CREDENTIALS=googleApplicationCredentials.json
+DEV_GUILD_ID=""
+DEV_CLIENT_I=""
